+++ onapp2vhi/ops/migrate.py
@@ -51,6 +51,7 @@ def migrate_impl(cfg: OnApp2VHIConfig,
                  placement='',
                  storage_policy='',
                  flavor='',
+                 cpu_hotplug=False
                  ):
     """
     Migrate all resources from OnApp to VHI:
@@ -81,6 +82,7 @@ def migrate_impl(cfg: OnApp2VHIConfig,
     :param cloud_init_install: project
     :param placement: placement "name" or "id"
     :param storage_policy: storage_policy "name"
+    :param cpu_hotplug: 'True' to enable, 'False' otherwise
     :return:
     """
     # Arrange
@@ -241,7 +243,8 @@ def migrate_impl(cfg: OnApp2VHIConfig,
                                    vdom=cfg.vhi_conf['vinfra_domain'],
                                    vm_properties=_vm_properties,
                                    vhi_obj=vhi,
-                                   placement=placement)
+                                   placement=placement,
+                                   cpu_hotplug=cpu_hotplug)
 
             vm_msg += (f'\t{_vm_number}. Migration Status = {result_vm}\n'
                        f'\t\t- IP "{_vm["ip_addr"]}"\n'